implementation using PurchaseDelegate
