+++ game-app/game-core/src/main/java/games/strategy/engine/data/GameState.java
@@ -3,6 +3,7 @@ package games.strategy.engine.data;
 import games.strategy.engine.data.properties.GameProperties;
 import games.strategy.triplea.delegate.AbstractMoveDelegate;
 import games.strategy.triplea.delegate.PoliticsDelegate;
+import games.strategy.triplea.delegate.PurchaseDelegate;
 import games.strategy.triplea.delegate.TechTracker;
 import games.strategy.triplea.delegate.TechnologyDelegate;
 import games.strategy.triplea.delegate.battle.BattleDelegate;
@@ -69,6 +70,8 @@ public interface GameState {
   BattleDelegate getBattleDelegate();
 
   PoliticsDelegate getPoliticsDelegate();
+
+  PurchaseDelegate getPurchaseDelegate();
 
   AbstractMoveDelegate getMoveDelegate();
