+++ game-app/game-core/src/main/java/games/strategy/triplea/delegate/PurchaseExtendedDelegateState.java
@@ -1,6 +1,8 @@
 package games.strategy.triplea.delegate;
 
+import games.strategy.engine.data.ProductionRule;
 import java.io.Serializable;
+import org.triplea.java.collections.IntegerMap;
 
 class PurchaseExtendedDelegateState implements Serializable {
   private static final long serialVersionUID = 2326864364534284490L;
@@ -8,4 +10,5 @@ class PurchaseExtendedDelegateState implements Serializable {
   Serializable superState;
   // add other variables here:
   boolean needToInitialize;
+  IntegerMap<ProductionRule> pendingProductionRules;
 }
